+++ vaadin-text-field.d.ts
@@ -1,2 +1,2 @@
 export * from './src/vaadin-text-field.js';
-export * from './@types/interfaces';
+export * from './src/interfaces';
